Update server socket's code rearrange
- swap socket.config.js with utils/socket.js
- change some var names for readability
- add some comments for readability
